requirements.txt
flask
ping3
speedtest-cli
gunicorn